Move user.py up a directory because it's the only file in a folder
